datas/requirements.txt
virtualenv venv
source /venv/Scripts/activate
djangorestframework
psycopg2-binary
requests